clara_3.0/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
jinja2==3.1.4
pydantic==2.7.1
PyMuPDF==1.24.2
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
langdetect==1.0.9
transformers==4.41.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.2
loguru==0.7.2
pyyaml==6.0
python-magic==0.4.27
pytest==7.4.2
